have split the module into multiple files.
This slightly affects load times but overall will improve the maintainability of the module going forwards.

From now on I am going to be focusing on implementing all functions available in the API fully featured. This will take time. To see the full list please consult with the csv that will appear with the next commit.